kernel-builder: Adds a WIP Pre-init logo
When using CONFIG_LOGO, this "pre-init" logo will be shown in the
center. This abuses the Linux mascot logo feature to help end-users
figure out without console logging whether the kernel is hung really
early, or after the framebuffer kicked-in. The initrd is expected to
replace the logo ASAP.

This is to be changed for something better integrated into the style
guide once written up.